+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `WithHttpClient` option support for all ingest client types (`New`, `NewStreaming`, `NewManaged`) to enable custom HTTP client configuration and instrumentation such as OpenTelemetry
+
 ## [1.0.3] - 2025-05-18
 
 ### Fixed